# noktajent/lib/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for Nokta
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /support/sessions:
    post:
      operationId: createSupportSession
      tags: [support]
      summary: Create or resume a support session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportSession"
  /support/sessions/{sessionId}/messages:
    get:
      operationId: getSupportMessages
      tags: [support]
      summary: Get messages for a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
        - name: after
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagesResponse"
    post:
      operationId: sendSupportMessage
      tags: [support]
      summary: Send a message in a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportMessage"
...
  /support/admin/sessions/{sessionId}/reply:
    post:
      operationId: sendAdminReply
      tags: [support]
      summary: Send a reply as support agent (admin)
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminReplyRequest"
      responses:
        "201":
          description: Reply sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportMessage"
  /support/admin/sessions/{sessionId}/status:
    patch:
      operationId: updateSessionStatus
      tags: [support]
      summary: Update session status (admin)
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportSession"
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task deleted
...
components:
  schemas:
...
    CreateSessionRequest:
      type: object
      properties:
        deviceId:
          type: string
        userName:
          type: string
      required: [deviceId]
    SupportSession:
      type: object
      properties:
        id:
          type: string
        deviceId:
          type: string
        userName:
          type: string
        status:
          type: string
          enum: [open, active, closed]
        createdAt:
          type: integer
        updatedAt:
          type: integer
      required: [id, deviceId, userName, status, createdAt, updatedAt]
    SupportMessage:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        role:
          type: string
          enum: [user, agent]
        text:
          type: string
        createdAt:
          type: integer
      required: [id, sessionId, role, text, createdAt]
    SendMessageRequest:
      type: object
      properties:
        text:
          type: string
      required: [text]
    MessagesResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/SupportMessage"
        session:
          $ref: "#/components/schemas/SupportSession"
      required: [messages, session]
...
    AdminReplyRequest:
      type: object
      properties:
        text:
          type: string
      required: [text]
    UpdateStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [open, active, closed]
      required: [status]
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, review, completed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        dueDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        userId:
          type: string
        projectId:
          type: integer
        tags:
          type: array
          items:
            type: string
        progress:
          type: integer
          minimum: 0
          maximum: 100
        checklist:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              text:
                type: string
              completed:
                type: boolean
            required: [id, text, completed]
      required: [id, title, status, priority, createdAt, updatedAt, userId]
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        dueDate:
          type: string
          format: date-time
        userId:
          type: string
        projectId:
          type: integer
        tags:
          type: array
          items:
            type: string
        checklist:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              text:
                type: string
              completed:
                type: boolean
            required: [id, text, completed]
      required: [title, userId]
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, review, completed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        dueDate:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
        progress:
          type: integer
          minimum: 0
          maximum: 100
        checklist:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              text:
                type: string
              completed:
                type: boolean
            required: [id, text, completed]
...